+++ src/widget/mypage/ui/StudyFormModal.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { FormProvider, useForm } from "react-hook-form";
+import { FormProvider } from "react-hook-form";
 
 import Button from "@/shared/components/atoms/Button";
 import Modal, {
@@ -11,34 +11,16 @@ import LabelInput from "@/shared/components/molecules/LabelInput";
 import LabelTextAreaInput from "@/shared/components/molecules/LabelTextAreaInput";
 
 import LabelSelectInput from "@/shared/components/molecules/LabelSelectInput";
-import { useMyJoinedStudyStore } from "@/shared/store/useMyJoinedStudyStore";
-import { useToast } from "@/shared/hooks/useToast";
-import { useQueryClient } from "@tanstack/react-query";
-import { useEffect, useMemo, useState } from "react";
 import LabelInputDateLocal from "@/shared/components/molecules/LabelDateTimeLocal";
 import { StudyScheduleInterface } from "@/shared/types/api/studies/detail";
 import { formatNowDate } from "@/shared/utils/date";
-import {
-  SchedulesQueryKey,
-  useDeleteStudySchedule,
-  usePostStudySchedule,
-  usePutStudySchedule,
-} from "@/entities/study/schedule/model/schedule.query";
 import { GetSchedulesResponse } from "@/entities/study/schedule/api/schedule.api.types";
+import useCalendarScheduleForm from "@/features/study/schedule/calendar-schedule-form/model/useCalendarScheduleForm";
 
 interface NoticeModalProps extends ModalProps {
   onClose: ModalOnClose;
   isModifyMode?: boolean;
   schedulesData?: GetSchedulesResponse[];
-}
-
-interface ScheduleData {
-  title: string;
-  description: string;
-  startDateTime: string;
-  endDateTime: string;
-  studyId?: number;
-  studyScheduleId?: number;
 }
 
 const StudyFormModal = ({
@@ -47,141 +29,27 @@ const StudyFormModal = ({
   schedulesData,
   ...props
 }: NoticeModalProps) => {
-  const queryClient = useQueryClient();
-  const methods = useForm<ScheduleData>({
-    defaultValues: {
-      title: "",
-      description: "",
-      startDateTime: "",
-      endDateTime: "",
-    },
-    mode: "onChange",
+  const {
+    methods,
+    handleSubmit,
+    onSubmit,
+    isCreating,
+    isUpdating,
+    isDeleting,
+    scheduleOption,
+    studyScheduleOption,
+    startDateTime,
+    studyScheduleId,
+    onClickDeleteBtn,
+    onClickCloseBtn,
+    selectedScheduleData,
+    isDelete,
+  } = useCalendarScheduleForm({
+    onClose,
+    isModifyMode,
+    schedulesData,
+    isOpen: props.isOpen,
   });
-  const { handleSubmit, reset, watch } = methods;
-  const { success, error } = useToast();
-  const studyId = Number(watch("studyId"));
-  const startDateTime = methods.watch("startDateTime");
-  const studyScheduleId = Number(watch("studyScheduleId"));
-  const [isDelete, setIsDelete] = useState<boolean>(false);
-
-  // 내가 참여중인 스터디 데이터
-  const myJoinedStudiesData = useMyJoinedStudyStore(
-    (state) => state.myJoinedStudiesData
-  );
-
-  // 스케줄 선택 옵션
-  const scheduleOption = useMemo(() => {
-    return myJoinedStudiesData?.map((item) => ({
-      label: item.title,
-      value: item.id,
-    }));
-  }, [myJoinedStudiesData]);
-
-  // 스케줄 일정 선택 옵션
-  const studyScheduleOption = useMemo(() => {
-    return schedulesData?.map((item) => ({
-      label: item.title,
-      value: item.studyScheduleId,
-    }));
-  }, [schedulesData]);
-
-  // 선택한 일정 상세 데이터 추출
-  const selectedScheduleData = useMemo(() => {
-    return schedulesData?.filter(
-      (item) => item.studyScheduleId === studyScheduleId
-    )[0];
-  }, [schedulesData, studyScheduleId]);
-
-  useEffect(() => {
-    // 수정 시 초기 데이터 셋팅
-    if (selectedScheduleData) {
-      methods.setValue("studyId", selectedScheduleData.studyId);
-      methods.setValue("title", selectedScheduleData.title);
-      methods.setValue("description", selectedScheduleData.description);
-      methods.setValue("startDateTime", selectedScheduleData.startDateTime);
-      methods.setValue("endDateTime", selectedScheduleData.endDateTime);
-    }
-  }, [selectedScheduleData, methods]);
-
-  // 일정 생성
-  const { mutate: createSchedule, isPending: isCreating } =
-    usePostStudySchedule(studyId, {
-      onSuccess: () => {
-        success("생성되었습니다.");
-        queryClient.invalidateQueries({
-          queryKey: SchedulesQueryKey,
-        });
-
-        reset();
-        onClose();
-      },
-      onError: (err) => {
-        error(String(err));
-      },
-    });
-
-  // 일정 수정
-  const { mutate: updateSchedule, isPending: isUpdating } = usePutStudySchedule(
-    studyId,
-    {
-      onSuccess: () => {
-        success("수정되었습니다.");
-        queryClient.invalidateQueries({
-          queryKey: SchedulesQueryKey,
-        });
-
-        reset();
-        onClose();
-      },
-      onError: (err) => {
-        error(String(err));
-      },
-    }
-  );
-
-  // 일정 삭제
-  const { mutate: deleteSchedule, isPending: isDeleting } =
-    useDeleteStudySchedule(studyId, {
-      onSuccess: () => {
-        success("삭제되었습니다.");
-        queryClient.invalidateQueries({
-          queryKey: SchedulesQueryKey,
-        });
-
-        setIsDelete(false);
-        reset();
-        onClose();
-      },
-      onError: (err) => {
-        error(String(err));
-      },
-    });
-
-  const onSubmit = (data: ScheduleData) => {
-    // 불필요한 필드 제거
-    delete data.studyId;
-
-    // API 호출
-    if (isDelete) {
-      // 삭제
-      deleteSchedule(studyScheduleId);
-    } else if (isModifyMode) {
-      // 수정
-      updateSchedule({ scheduleId: studyScheduleId, data: data });
-    } else {
-      // 셍성
-      createSchedule(data);
-    }
-  };
-
-  const onClickDeleteBtn = () => {
-    setIsDelete(true);
-  };
-
-  const onClickCloseBtn = () => {
-    reset();
-    onClose();
-  };
 
   return (
     <FormProvider {...methods}>
